+++ motor/rule_tests/owasp_tests/no_credentials_in_url_test.go
@@ -1,0 +1,91 @@
+package tests
+
+import (
+	"testing"
+
+	"github.com/daveshanley/vacuum/model"
+	"github.com/daveshanley/vacuum/motor"
+	"github.com/daveshanley/vacuum/rulesets"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestRuleSet_GetOWASPRuleSecurityCredentialsDetected_Success(t *testing.T) {
+
+	yml := `openapi: "3.1.0"
+paths:
+  /foo/{id}/:
+    get:
+      description: "get"
+      parameters:
+        - name: id
+          in: path
+          required: true
+        - name: filter
+          in: query
+          required: true`
+
+	rules := make(map[string]*model.Rule)
+	rules["here"] = rulesets.GetOWASPRuleSecurityCredentialsDetected() // TODO
+
+	rs := &rulesets.RuleSet{
+		Rules: rules,
+	}
+
+	rse := &motor.RuleSetExecution{
+		RuleSet: rs,
+		Spec:    []byte(yml),
+	}
+	results := motor.ApplyRulesToRuleSet(rse)
+	assert.Len(t, results.Results, 0)
+}
+
+func TestRuleSet_GetOWASPRuleSecurityCredentialsDetected_Error(t *testing.T) {
+
+	yml := `openapi: "3.1.0"
+paths:
+  /foo/{id}/:
+    get:
+      description: "get"
+      parameters:
+        - name: client_secret
+          in: query
+          required: true
+        - name: token
+          in: query
+          required: true
+        - name: refresh_token
+          in: query
+          required: true
+        - name: id_token
+          in: query
+          required: true
+        - name: password
+          in: query
+          required: true
+        - name: secret
+          in: query
+          required: true
+        - name: apikey
+          in: query
+          required: true
+        - name: apikey
+          in: path
+          required: true
+        - name: API-KEY
+          in: query
+          required: true`
+
+	rules := make(map[string]*model.Rule)
+	rules["here"] = rulesets.GetOWASPRuleSecurityCredentialsDetected() // TODO
+
+	rs := &rulesets.RuleSet{
+		Rules: rules,
+	}
+
+	rse := &motor.RuleSetExecution{
+		RuleSet: rs,
+		Spec:    []byte(yml),
+	}
+	results := motor.ApplyRulesToRuleSet(rse)
+	assert.Len(t, results.Results, 9)
+}
