+++ motor/rule_tests/owasp_tests/no_numeric_ids_test.go
@@ -1,0 +1,84 @@
+package tests
+
+import (
+	"testing"
+
+	"github.com/daveshanley/vacuum/model"
+	"github.com/daveshanley/vacuum/motor"
+	"github.com/daveshanley/vacuum/rulesets"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestRuleSet_TestGetOWASPRuleAPIRuleNoNumericIDs_Success(t *testing.T) {
+
+	yml := `openapi: "3.1.0"
+paths:
+  /foo/{id}/:
+    get:
+      description: "get"
+      parameters:
+        - name: id
+          in: path
+          schema:
+            type: string
+            format: uuid`
+
+	rules := make(map[string]*model.Rule)
+	rules["here"] = rulesets.GetOWASPRuleAPIRuleNoNumericIDs()
+
+	rs := &rulesets.RuleSet{
+		Rules: rules,
+	}
+
+	rse := &motor.RuleSetExecution{
+		RuleSet: rs,
+		Spec:    []byte(yml),
+	}
+	results := motor.ApplyRulesToRuleSet(rse)
+	assert.Len(t, results.Results, 0)
+}
+
+func TestRuleSet_TestGetOWASPRuleAPIRuleNoNumericIDs_Error(t *testing.T) {
+
+	yml := `openapi: "3.1.0"
+paths:
+  /foo/{id}/:
+    get:
+      description: "get"
+      parameters:
+        - name: id
+          in: path
+          schema:
+            type: integer
+        - name: notanid
+          in: path
+          schema:
+            type: integer
+        - name: underscore_id
+          in: path
+          schema:
+            type: integer
+        - name: hyphen-id
+          in: path
+          schema:
+            type: integer
+            format: int32
+        - name: camelId
+          in: path
+          schema:
+            type: integer`
+
+	rules := make(map[string]*model.Rule)
+	rules["here"] = rulesets.GetOWASPRuleAPIRuleNoNumericIDs()
+
+	rs := &rulesets.RuleSet{
+		Rules: rules,
+	}
+
+	rse := &motor.RuleSetExecution{
+		RuleSet: rs,
+		Spec:    []byte(yml),
+	}
+	results := motor.ApplyRulesToRuleSet(rse)
+	assert.Len(t, results.Results, 5) // in spectral, this outputs 4 errors // TO DISCUSS
+}
